Clock tests in own file.
