+++ sdk/core/src/test/java/com/klaviyo/coresdk/model/ProfileTest.kt
@@ -1,0 +1,3 @@
+package com.klaviyo.coresdk.model
+
+class ProfileTest
